# doc/api/api-swagger.yml
openapi: 3.0.0
info:
  description: |
    This is the Prophecy Api documentation
  version: 1.0.0
  title: "Prophecy Api"
...
paths:
...
  /account/sign-in:
    post:
      summary: Enables the user to sign in
      description: "**Request body:**\n\nThe *username* field can either be the username or the email address."
      tags:
        - account
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/signInCredentials'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/connectionResponse'
        400:
          description: Bad Request
        401:
          description: |
            Unauthorized 
            
            The email address and/or the password is/are invalid
        500:
          description: Inernal Server Error (unable to connect to the database)
  /account/sign-up:
    post:
      summary: Enables the user to create an account
      description: "Sends a email confirmation link and doesn't sign in the user"
      tags:
        - account
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/signUpCredentials'
      responses:
        201:
          description: Account created
        400:
          description: Bad Request
        409:
          description: Conflict (the email address is already in use)
        500:
          description: Internal Server Error (unable to connect to the database)
  /account/sign-out:
    post:
      summary: Enables the user to logout
      description: ""
      tags:
        - account
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parametersSchemes/authorization'
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        401:
          $ref: '#/components/schemas/unauthorizedResponse'
        500:
          description: Internal Server Error
...
  /account/settings/update-password:
    put:
      summary: Updates the user's password
      description: ""
      tags:
        - account
        - settings
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parametersSchemes/authorization'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/password'
      responses:
        200:
          description: OK
        400:
          description: Bad request
        401:
          $ref: '#/components/schemas/unauthorizedResponse'
        500:
          description: Internal Server Error
  /account/settings/update-username:
    put:
      summary: Updates the user's username
      description: ""
      tags:
        - account
        - settings
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parametersSchemes/authorization'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/username'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/username'
        400:
          description: Bad Request
        401:
          $ref: '#/components/schemas/unauthorizedResponse'
        500:
          description: Internal Server Error (unable to connect to the database)
  /account/settings/update-email-address:
    put:
      summary: Updates the user's email address
      description: ""
      tags:
        - account
        - settings
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parametersSchemes/authorization'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/email'
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        401:
          $ref: '#/components/schemas/unauthorizedResponse'
        409:
          description: Conflict (email address already in use)
        422:
          description: Unporcessable Entity (invalid email address)
        500:
          description: Internal Server Error (unable to connect to the database)

  # delete
  /account/settings/delete-account:
    delete:
      summary: Deletes the user's account
      description: ""
      tags:
        - account
        - settings
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parametersSchemes/authorization'
      responses:
        200:
          description: OK
        400:
          description: Bad Request
        401:
          $ref: '#/components/schemas/unauthorizedResponse'
        500:
          description: Internal Server Error

  # PROFILE
...
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/Victoire-Rabeau/Prophecy/1.0.0
  - description: The local server
    url: localhost
    variables:
      basePath:
        default: /api
components:
  schemas:
    password:
      type: string
      description: The user's password (length up to 72 bytes)
      example: password
    username:
      type: string
      description: The user's username
      example: username
    email:
      type: string
      description: The user's email address
      example: email@address.com
...
    signInCredentials:
      description: Informations required to sign in
      type: object
      properties:
        username:
          description: "username or email address"
        password:
          $ref: '#/components/schemas/password'
      required:
        - email
        - password
    signUpCredentials:
      description: Informations required to sign up
      type: object
      properties:
        username:
          $ref: '#/components/schemas/username'
        email:
          $ref: '#/components/schemas/email'
        password:
          $ref: '#/components/schemas/password'
      required:
        - username
        - email
        - password
    connectionResponse:
      description: The data sent on connection
      properties:
        access_token:
          type: string
          description: The required token
        username:
          type: string
          description: The user's id
          format: string
          example: "username"
      required:
        - accessToken
        - userId
...
    unauthorizedResponse:
      description: |
        Unauthorized
        
        The token is not valid
...
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parametersSchemes:
    authorization:
      in: header
      name: Authorization
      description: The authorization bearer token
      required: true
      schema:
        $ref: '#/components/securitySchemes/bearerAuth'
...